+++ src/TileSet.cpp
@@ -23,7 +23,7 @@ TileSet::TileSet(int tileWidth, int tileHeight, std::string filename) : tileWidt
 void TileSet::RenderTile(unsigned index, int x, int y)
 {
   // Ensure valid index
-  Assert(index >= 0 && index < rows * columns, "Invalid index");
+  Assert(index >= 0 && (int)index < rows * columns, "Invalid index");
 
   // Get to which row this tile belongs
   int tileRow = index / columns;
